requirements.txt
colorama==0.4.6
python-nmap==0.7.1